Fix various bugs in checksumming code
This commit cleans up a lot of things in the checksumming code, among
those:
- Proper types are used throughout the code (to avoid unintended
  overflows/sign extensions/confusion while reading the code).
- Documentation of the methods is fixed where it was wrong/outdated.
- Better error logging is done for some situations.
- Bug fix: the span between the end of the block and the data being
  inserted was not cleaned when the inserted data is placed past the
  end.
- Bug fix: incomplete file reads were not handled correctly, which could
  lead to spurious crashes being attributed to data corruption.
- Bug fix: it was impossible to request to read more data than is
  available in the last block, expecting it to return the data that is
  available.
- Methods are made more readable, so that the code could be more easily
  improved; currently the code is very inefficient, but all the
  optimizations are left out of this commit for now.

It also adds a better set of tests to go with the code, however these
tests do not cover the checksumming functionality completely.

To run the tests you'd have to do the following (for now):

  cd mysys
  make test_checksum_io
  ./test_checksum_io

+++ mysys/my_pread.c
@@ -60,7 +60,7 @@ size_t my_pread(File Filedes, uchar *Buffer, size_t Count, my_off_t offset,
 #endif
 #ifndef HAVE_PREAD
     pthread_mutex_lock(&my_file_info[Filedes].mutex);
-    readbytes= (uint) -1;
+    readbytes= (size_t) -1;
     error= (lseek(Filedes, offset, MY_SEEK_SET) == (my_off_t) -1 ||
 	    (readbytes= read(Filedes, Buffer, (uint) Count)) != Count);
     pthread_mutex_unlock(&my_file_info[Filedes].mutex);
